[Fix] Design update
